requirements.txt
beautifulsoup4==4.14.3
lxml==6.0.2
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.46
streamlit==1.42.0
plotly==6.0.0